refactor: expose search criteria from common only
