+++ datasketches/src/req/sketch.rs
@@ -22,13 +22,13 @@ use crate::codec::SketchSlice;
 use crate::codec::assert::insufficient_data;
 use crate::codec::family::Family;
 use crate::common::NumStdDev;
+use crate::common::SearchCriteria;
 use crate::error::Error;
 use crate::req::DEFAULT_K;
 use crate::req::INITIAL_SECTIONS_PER_COMPACTOR;
 use crate::req::MAX_K;
 use crate::req::MIN_K;
 use crate::req::RankAccuracy;
-use crate::req::SearchCriteria;
 use crate::req::compactor::Compactor;
 use crate::req::iter::ReqSketchIterator;
 use crate::req::serialization::FLAG_IS_EMPTY;
