+++ tests-integration/tests/req_test/accuracy.rs
@@ -17,9 +17,9 @@
 
 //! End-to-end accuracy checks for ReqSketch.
 
+use datasketches::common::SearchCriteria;
 use datasketches::error::Error;
 use datasketches::req::ReqSketch;
-use datasketches::req::SearchCriteria;
 use googletest::assert_that;
 use googletest::prelude::le;
